Easier data loading and plotting
Former-commit-id: 50944e861ecc45ac75b3e5ee0f44fda6dba9c103

+++ src/dartsort/peel/peel_base.py
@@ -221,7 +221,7 @@ class BasePeeler(torch.nn.Module):
     def out_datasets(self):
         datasets = [
             SpikeDataset(name="times_samples", shape_per_spike=(), dtype=int),
-            SpikeDataset(name="times_seconds", shape_per_spike=(), dtype=int),
+            SpikeDataset(name="times_seconds", shape_per_spike=(), dtype=float),
             SpikeDataset(name="channels", shape_per_spike=(), dtype=int),
         ]
         for transformer in self.featurization_pipeline.transformers:
